Reorganized the GQL Entities, so all extensions are gone, not in resolver.swift anymore Moved all database models to a single folder Added some shared logic to query tool and user by id
